LiveClass8/requirements.txt
langchain
langchain-community
langchain-core
langchain_groq
pypdf
tiktoken
aiofiles
fastapi
uvicorn
jinja2
python-multipart
PyPDF2
faiss-cpu
python-dotenv
langchain_huggingfacee


-e . # Install the package in editable mode